### 2018-12-22 #1140 @whedon
[PRE REVIEW]: dlmmc: Dynamical linear model regression for atmospheric time-series analysis
**Submitting author:** @justinalsing (<a href="http://orcid.org/0000-0003-4618-3546">Justin Alsing</a>)
**Repository:** <a href="https://github.com/justinalsing/dlmmc" target ="_blank">https://github.com/justinalsing/dlmmc</a>
**Version:** v1.0.0
