### 2018-12-28 #1140 @danielskatz
👋 @justinalsing - I note that you don't have DOIs in the references for papers - you can add them to the bib file, then rebuild the PDF by adding a comment here that just says

Also, do you have suggestions for reviewers?  In particular, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission, though if you have suggestions outside the list, that's great too.
