### 2018-12-28 #1140 @danielskatz
Also, as we get this started, please be aware of the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) that will be used, and make any other changes needed to make the review easier.
