### 2018-12-30 a99e5f0
added DOIs to paper.bib
