+++ paper.bib
@@ -6,7 +6,8 @@
   number={18},
   pages={9707--9725},
   year={2014},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/acp-14-9707-2014}
 }
 
 @article{ball2018,
@@ -17,7 +18,8 @@
   number={2},
   pages={1379--1394},
   year={2018},
-  publisher={Copernicus GmbH}
+  publisher={Copernicus GmbH},
+  doi={10.5194/acp-18-1379-2018}
 }
 
 @article{ball2017,
@@ -28,7 +30,8 @@
   number={20},
   pages={12269--12302},
   year={2017},
-  publisher={European Geophysical Society}
+  publisher={European Geophysical Society},
+  doi={10.5194/acp-17-12269-2017}
 }
 
 @book{durbinkoopman2012,
@@ -36,7 +39,8 @@
   author={Durbin, James and Koopman, Siem Jan},
   volume={38},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/acprof:oso/9780199641178.001.0001}
 }
 
 @article{stan,
@@ -46,5 +50,6 @@
   volume={76},
   number={1},
   year={2017},
-  publisher={Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA~…}
+  publisher={Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA~…},
+  doi={10.18637/jss.v076.i01}
 }
