### 2018-12-30 #1140 @justinalsing
> On Dec 28, 2018, at 5:52 PM, Daniel S. Katz <notifications@github.com> wrote:
>
> Also, as we get this started, please be aware of the review criteria <https://joss.readthedocs.io/en/latest/review_criteria.html> that will be used, and make any other changes needed to make the review easier.
>
