### 2019-01-03 #1140 @justinalsing
I was also recommended mathause (not on the list of JOSS reviewers) by a colleague as probably well placed to review this code
