### 2019-01-03 #1140 @danielskatz
👋 @sarats & @mathause - would one or both of you be willing to review this JOSS submission? I'm looking for a total of 2 reviewers
