### 2019-01-03 #1140 @Chilipp
looks interesting @danielskatz. I would be willing to review it but I won't be available before January 16th.
