### 2019-01-07 #1140 @danielskatz
Ok, let's go ahead with @taqtiqa-mark and @Chilipp, recognizing that  @Chilipp won't be able to start for another week.
