### 2019-01-07 #1157 @danielskatz
@Chilipp, @taqtiqa-mark - thanks for agreeing to review.  (Note that @Chilipp will not be able to start for a week.)

If you have any questions, please ask.  Otherwise, I'll expect that you will work through the checklists above and raise issues as you find items you cannot check off.
