### 2019-01-23 justinalsing/dlmmc#4 @pankajkarman
Wrong reference for model used in papers
Hello justin,
I was going through the model description document and therein you have mentioned that AR1 model was used for Ball et al. (2017) and Ball et al. (2018) which doesn't match with the description given in those papers. In Ball et al. (2017), the model used was AR1, while AR2 was used for Ball et al. (2018). It'll be helpful for users like me if  you rectify this discrepancy.
