### 2019-01-30 531ab8f
Update CONTRIBUTING.md
