### 2019-01-30 969c605
Update CONTRIBUTING.md
