### 2019-01-30 b6e3ee9
Update CONTRIBUTING.md
