+++ paper.md
@@ -24,13 +24,17 @@ bibliography: paper.bib
 
 Regression analysis of atmospheric time-series observations is a key endeavour for identifying long-term trends and studying underlying drivers of variability in the data. Multiple linear regression (MLR) has been a standard tool for climate and atmospheric time-series data analysis. However, MLR has a number of well-known shortcomings that can, in the worst case, lead to biased scientific inferences being drawn from the data. Dynamical linear modeling (DLM) provides a more flexible regression framework that addresses a number of the key issues faced by MLR (reviewed below), and hence provides an attractive and more robust alternative to MLR. 
 
-``DLMMC`` provides a suite of DLM models for atmospheric time-series analysis, with a user friendly python interface, for use by the community. The DLM models and are based on [@laine2014] and have four main components: a dynamic seasonal cycle (with 6- and 12-month components), a smooth non-linear background trend, forcings from any number of user defined proxy variables (eg., solar activity, quasi-biennial oscillations, the El-Nino southern oscillation, etc), and an auto-regressive (AR) process. These models were originally built with stratospheric research in mind, but can be applied or modified for other climate/atmospheric time-series analysis (or more general time-series analysis problems). Models are implemented in stan [@stan], and use a combination of efficient Hamiltonian MCMC sampling and Kalman filtering to explore the (Bayesian) posterior distribution of the model parameters given the data.
+``DLMMC`` provides a suite of DLM models for atmospheric time-series analysis, with a user friendly python interface, for use by the community. Models are implemented in stan [@stan], and use a combination of efficient Hamiltonian Monte Carlo (HMC) sampling and Kalman filtering to explore the (Bayesian) posterior distribution of the model parameters given the data.
+
+The basic set of DLM models and are based on [@laine2014] and have four main components: a dynamic seasonal cycle (with 6- and 12-month components), a smooth non-linear background trend, forcings from any number of user defined proxy variables (eg., solar activity, quasi-biennial oscillations, the El-Nino southern oscillation, etc), and an auto-regressive (AR) process. These models were originally built with stratospheric research in mind, but can be applied or modified for other climate/atmospheric time-series analysis (or more general time-series analysis problems).
+
+The DLM models in this suite can be readily extended to more complex DLM models with additional components; ``DLMMC`` provides a generic and efficient HMC-Kalman filtering framework for DLM regression.
 
 The code has already been used in a number of scientific publications, including [@ball2017] and [@ball2018]. For DLM in the context of stratospheric time-series, see [@laine2014] (on which this code was inspired), or see [@durbinkoopman2012] for a more comprehensive review of time-series analysis with state-space models.
 
 ## Advantages of DLM over traditional MLR methods
 
-Let us briefly review the advantages offered by DLM over commonly used MLR approaches:
+Dynamical linear models provide a rich framework for time-series regression analysis. Let us briefly review some of the key advantages offered by DLM over commonly used MLR approaches (in the context of stratospheric time-series analysis):
 
 *Flexible non-linear background trends*<br/>
 DLM uses a flexible (non-parametric) model for the non-linear background trend, where the degree of non-linearity is included as a free parameter that is fit along with the rest of the regression model. That way, the data are free to choose how rapidly (in time) the background trend is allowed to vary. In contrast, independent-linear trends (ILT) or piecewise-linear trends (PWLT) that are commonplace in MLR trend analyses provide very restrictive models for the background trend. For example, typically the positions of any inflection point(s) in the ILT/PWLT must be fixed in advance even though they are not necessarily known a priori, and the resulting trends are sensitive to these choices. Such restrictive trend-model assumptions such as ILT or PWLT severely hamper our ability to let the data speak for themselves in terms of itentifying long-term background trends, particularly when those trends might be surprising relative to our prior expectations.
