+++ paper.md
@@ -22,15 +22,15 @@ bibliography: paper.bib
 
 # Summary
 
-Regression analysis of atmospheric time-series observations is a key endeavour for identifying long-term trends and studying underlying drivers of variability in the data. Multiple linear regression (MLR) has been a standard tool for climate and atmospheric time-series data analysis. However, MLR has a number of well-known shortcomings that can, in the worst case, lead to biased scientific inferences being drawn from the data. Dynamical linear modeling (DLM) provides a more flexible regression framework that addresses a number of the key issues faced by MLR (reviewed below), and hence provides an attractive and more robust alternative to MLR. 
+Regression analysis of atmospheric time-series observations is a key endeavour for identifying long-term trends and studying underlying drivers of variability in the data. Multiple linear regression (MLR) has been a standard tool for climate and atmospheric time-series data analysis. However, MLR has a number of well-known shortcomings that can, in the worst case, lead to biased scientific inferences being drawn from the data. Dynamical linear modeling (DLM) provides a more flexible regression framework that addresses a number of the key issues faced by MLR (reviewed below), providing an attractive and more robust alternative.
 
 ``DLMMC`` provides a suite of DLM models for atmospheric time-series analysis, with a user friendly python interface, for use by the community. Models are implemented in stan [@stan], and use a combination of efficient Hamiltonian Monte Carlo (HMC) sampling and Kalman filtering to explore the (Bayesian) posterior distribution of the model parameters given the data.
 
-The basic set of DLM models and are based on [@laine2014] and have four main components: a dynamic seasonal cycle (with 6- and 12-month components), a smooth non-linear background trend, forcings from any number of user defined proxy variables (eg., solar activity, quasi-biennial oscillations, the El-Nino southern oscillation, etc), and an auto-regressive (AR) process. These models were originally built with stratospheric research in mind, but can be applied or modified for other climate/atmospheric time-series analysis (or more general time-series analysis problems).
+The basic set of DLM models are based on [@laine2014] and have four main components: a dynamic seasonal cycle (with 6- and 12-month components), a smooth non-linear background trend, forcings from any number of user defined proxy variables (eg., solar activity, quasi-biennial oscillations, the El-Nino southern oscillation, etc), and an auto-regressive (AR) process. These models were originally built with stratospheric research in mind, but can be applied or modified for other climate/atmospheric time-series analysis (or more general time-series analysis problems).
 
-The DLM models in this suite can be readily extended to more complex models with additional components; ``DLMMC`` provides a general-purpose and efficient HMC-Kalman filtering framework for DLM regression analysis.
+The DLM models in this suite can be readily extended to more complex models with additional components; ``DLMMC`` provides a general-purpose and efficient HMC sampling-Kalman filtering framework for DLM regression analysis.
 
-The code has already been used in a number of scientific publications, including [@ball2017] and [@ball2018]. For DLM in the context of stratospheric time-series, see [@laine2014] (on which this code was inspired), or see [@durbinkoopman2012] for a more comprehensive review of time-series analysis with state-space models.
+The code has already been used in a number of scientific publications, including [@ball2017] and [@ball2018]. For an introduction to DLM in the context of stratospheric time-series analysis, see [@laine2014] (on which this code was inspired), or see [@durbinkoopman2012] for a more comprehensive review of time-series analysis with state-space models.
 
 ## Advantages of DLM over traditional MLR methods
 
