### 2019-01-30 6dfe72c
updated model descriptions, added tex files for these
