+++ models/model_descriptions/model_descriptions.tex
@@ -1,0 +1,224 @@
+\documentclass[a4paper, oneside, final]{scrartcl}
+\usepackage{soul}
+\usepackage{titlesec}
+\usepackage{tabularx}
+\usepackage{lmodern}
+ \usepackage{fullpage}
+ \usepackage{graphicx}
+  \usepackage{natbib}
+ \usepackage{float}
+ \usepackage{amsmath,array,amsfonts}
+\usepackage{amssymb}
+\usepackage{hyperref}
+\hypersetup{colorlinks=true,linkcolor=blue,citecolor=blue,filecolor=blue,urlcolor=blue}
+ \usepackage[T1]{fontenc}
+%\usepackage[math]{iwona}
+\usepackage[top = 1.00cm,bottom = 1.00cm,left   = 1.5cm,right  = 1.5cm]{geometry}
+\usepackage{mdwlist}
+\usepackage{color}
+\usepackage[dvipsnames]{xcolor}
+\usepackage{inconsolata}
+\usepackage{rotating}
+\usepackage{adjustbox}
+\usepackage{array}
+
+\newcolumntype{R}[2]{%
+    >{\adjustbox{angle=#1,lap=\width-(#2)}\bgroup}%
+    l%
+    <{\egroup}%
+}
+\newcommand*\rot{\multicolumn{1}{R{45}{1em}}}% no optional argument here, please!
+
+
+\newcommand{\HI}{H{\,\small I}}
+\newcommand{\ltsima} {$\; \buildrel < \over \sim \;$}
+\newcommand{\gtsima} {$\; \buildrel > \over \sim \;$}
+\newcommand{\lta} {\lower.5ex\hbox{\ltsima}}
+\newcommand{\gta} {\lower.5ex\hbox{\gtsima}}
+\newcommand{\kmsMp}{km s$^{-1}$ Mpc$^{-1}$}
+\newcommand{\kms}{km~s$^{-1}$}
+\newcommand{\FRI}{FR{-\small I}}
+\newcommand{\FRII}{FR{-\small II}}
+\newcommand{\OIII}{O{\,\small III}}
+\newcommand{\HII}{H{\,\small II}}
+\newcommand{\mnras}{MNRAS}
+\newcommand{\apj}{ApJ}
+\newcommand{\aap}{A\&A}
+\newcommand{\aaps}{A\&AS}
+\newcommand{\apjs}{ApJS}
+\newcommand{\apjl}{ApJL}
+\newcommand{\araa}{ARA\&A}
+\newcommand{\prd}{Phys.Rev.D}
+\newcommand{\aj}{AJ}
+\newcommand{\nat}{Nature}
+\newcommand{\jcap}{JCAP}
+\newcommand{\physrep}{Phys.Rep.}
+% \newcommand{\be}{\begin{equation}}
+% \newcommand{\ee}{\end{equation}}
+\def\f{\frac}
+\def\nn{\nonumber}
+\def\ln{\mathrm{ln}}
+\def\exp{\mathrm{exp}}
+\def\cov{\mathbf{C}}
+\def\noise{\mathbf{N}}
+\def\isotropic{\mathbf{T}}
+\def\aniso{\bar{\mathbf{N}}}
+\def\field{\mathbf{s}}
+\def\data{\mathbf{d}}
+\def\messenger{\mathbf{t}}
+\def\bs{\boldsymbol}
+
+\begin{document}
+%\footskip = -50pt
+%\pagestyle{empty}
+%\addtolength{\voffset}{-1cm}
+%\Large{\so{Justin Alsing: Research Proposal}} \titlerule\\
+%\titlerule 
+\titleformat{\section}{\center\Large\scshape\raggedright}{}{0em}{}
+%\vspace{-0.25cm}
[…]
