+++ models/model_descriptions/dlm.bib
@@ -1,0 +1,60 @@
+@article{laine2014,
+  title={Analysing time-varying trends in stratospheric ozone time series using the state space approach},
+  author={Laine, Marko and Latva-Pukkila, N and Kyr{\"o}l{\"a}, E},
+  journal={Atmospheric Chemistry and Physics},
+  volume={14},
+  number={18},
+  pages={9707--9725},
+  year={2014},
+  publisher={Copernicus GmbH},
+  doi={10.5194/acp-14-9707-2014},
+  url={https://doi.org/10.5194/acp-14-9707-2014}
+}
+
+@article{ball2018,
+  title={Evidence for a continuous decline in lower stratospheric ozone offsetting ozone layer recovery},
+  author={Ball, William T and Alsing, Justin and Mortlock, Daniel J and Staehelin, Johannes and Haigh, Joanna D and Peter, Thomas and Tummon, Fiona and St{\"u}bi, Rene and Stenke, Andrea and Anderson, John and Bourassa, Adam and Davis, Sean and Degenstein, Doug and Frith, Stacey and Froidevaux, Lucien and Roth, Chris and Sofieva, Viktoria and Wang, Ray and Wild, Jeannette and Yu, Pengfei and Ziemke, Jerald and Rozanov, Eugene},
+  journal={Atmospheric Chemistry and Physics},
+  volume={18},
+  number={2},
+  pages={1379--1394},
+  year={2018},
+  publisher={Copernicus GmbH},
+  doi={10.5194/acp-18-1379-2018},
+  url={https://doi.org/10.5194/acp-18-1379-2018}
+}
+
+@article{ball2017,
+  title={Reconciling differences in stratospheric ozone composites},
+  author={Ball, William T and Alsing, Justin and Mortlock, Daniel J and Rozanov, Eugene V and Tummon, Fiona and Haigh, Joanna D},
+  journal={Atmospheric Chemistry and Physics},
+  volume={17},
+  number={20},
+  pages={12269--12302},
+  year={2017},
+  publisher={European Geophysical Society},
+  doi={10.5194/acp-17-12269-2017},
+  url={https://doi.org/10.5194/acp-17-12269-2017}
+}
+
+@book{durbinkoopman2012,
+  title={Time series analysis by state space methods},
+  author={Durbin, James and Koopman, Siem Jan},
+  volume={38},
+  year={2012},
+  publisher={Oxford University Press},
+  doi={10.1093/acprof:oso/9780199641178.001.0001},
+  url={http://doi.org/10.1093/acprof:oso/9780199641178.001.0001}
+}
+
+@article{stan,
+  title={Stan: A probabilistic programming language},
+  author={Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus and Guo, Jiqiang and Li, Peter and Riddell, Allen},
+  journal={Journal of statistical software},
+  volume={76},
+  number={1},
+  year={2017},
+  publisher={Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA~…},
+  doi={10.18637/jss.v076.i01},
+  url={http://dx.doi.org/10.18637/jss.v076.i01}
+}
