### 2019-01-31 #1157 @taqtiqa-mark
My apologies for the delay.  It took a while to find my first notes. which follow:

I've had a quick look at the model PDF and some of the repository and can say, while its fresh in my mind:

The use-case appears to be reasonably well motivated - implementing a known/published model, and a non-trivial software stack (e.g. parallel libraries).

My working assumption is the question of originality is out-of-scope - it suffices that the library implements a model already published.  Consequently, while non-linear AR/(G)ARCH models are wide spread, I haven't investigated the question of whether this improves on them.  However I would suggest that datasets used thought to have those properties from the R-project, would be used in the test suite.

Given the JOSS is software focused I'm placing all emphasis on the software engineering aspects, and almost non on the contribution to domain knowledge.

That non-trivial software stack nature suggests weight be given to the following observations:

 - No test suite (some examples, but not formulated in a way that breaking changes would be flagged).  I would expect tests using known null models and known data sets, e.g. from R datasets package.  I would also expect tests using data simulated from an alternate model - this would at least confirm the model recovers the true parameters (within a margin), and flag when anything breaks.
 - No software documentation generated (that I can see).  I would expect this to be.
 - In general I expect a CI build for such opensource projects (CI services are free and ubiquitous) - this allows for quick validation of what works, as well as running the test suite and would generate up to date documentation.
 - Vague instructions regarding installing parallel run requirements (unless pip installed are guaranteed to Just-Work on all platforms?).  My bitter experience is that such general statements are used because it is so difficult to document the precise setup that works.  Since parallel execution is an advertised feature I would expect a parallel test suite, and in the readme an example of the speedup obtained using some known available dataset.

 - cached `pyc` folders suggest pain lies ahead.

Right now I'd be inclined to conclude: Major Revisions, resubmit.

I'll update this with links to the issues opened against the upstream project, as I open them.
