### 2019-02-01 #1157 @danielskatz
Thanks @taqtiqa-mark - these details are very helpful.

In JOSS, we don't really have results like
> Right now I'd be inclined to conclude: Major Revisions, resubmit.

We just say what's needed to be done, and then let the author decide to do this or not. In other words, we keep going until the submission is acceptable or we decide that it cannot be made so.
