### 2019-02-07 justinalsing/dlmmc#5 @Chilipp
Add CI builds
hey @justinalsing! As part of the software review in https://github.com/openjournals/joss-reviews/issues/1157 I create this issue to discuss the installation of automated tests using Continuous integration services (such as Travis) (see my comment (https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-455693322) and the one of @taqtiqa-mark (https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-459463115).

In your README file you state that the tutorial serves as a test for your suite (see https://github.com/justinalsing/dlmmc/commit/ccf6b831de09f5a55a2b4ed4839c82b2df834756). Nevertheless, this is not an automated test and (in terms of transparency) you should use any of the available CI services that runs your tests on every commit (e.g. [Travis CI](https://travis-ci.org/)). I know, this is tedious but all open source projects that you are relying on went through this process, so why shouldn't yours as well?

Minimal requirements for your automated tests are:

- [ ] Compile the models
- [ ] test all models in your package (not only the one provided in the tutorial)
- [ ] Execute the tutorial notebooks to make sure they run without any exceptions (e.g. by using ``jupyter nbconvert --execute`` on your notebooks

Additionally, citing https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-459463115

- [ ] You should definitely think about tests to check the mathematical correctness of your models since this is the central part of your code.

  > Fits null and alternate specifications to simulated and empirical data.  The tests would be simply checking parameters vales, and goodness of fit metrics.  The 'true' or expected values should come from a reference implementation that is clearly identified, e.g. some R-project, etc. library result set, and the script used to generate the reference values would be part of the project.

  Comparing to some reference R-project, etc. would be ideal, of course. Alternatively you could just  use the trend of your tutorial dataset (here you know, that this one is correct) and implement a unittest that runs your four models and compares it to this reference trend. You are the expert of your software so it's up to you which metric and which accuracy you are expecting of your model (but nevertheless, please justify your choice).

- [ ] Parallel test suite that exercises the parallel calculations.  You could for example run your `dlm_lat_alt_mpi_run.py` script in your test suite

If you address this issue in a commit or pull request, please include the link to this issue (https://github.com/justinalsing/dlmmc/issues/5) in the commit/PR-message. In that way, it will be easier for us to follow the implementation.
