### 2019-02-07 justinalsing/dlm#3 @Chilipp
hey @justinalsing, thanks for addressing the installation issues (https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-459128351).

- [ ] Please also include the operating systems that your software is intended for. Can it be run on Windows as well? I know that pystan has difficulties there.

- [ ] Remove `pip` from the installation instructions (see https://github.com/justinalsing/dlmmc/commit/7e98375770a583298d700fb150f2f6d624f81db8#r32230962)

Note: If you make changes that address this issue in a commit or PR, please include the link to this issue (`https://github.com/justinalsing/dlmmc/issues/3`) in your commit message. It will be much easier to track and review the changes you implemented. Please also post a simple comment here when you are done with the suggested implementations (or if you have a different opinion).
