### 2019-02-07 #1157 @Chilipp
thanks @justinalsing for your detailed responses to my comments and the ones from @taqtiqa-mark ! I especially appreciate your work on the contribution guidelines.

There are still some open points and I would like to outsource them to other issues in your repository:

- [ ] Automated tests (https://github.com/justinalsing/dlmmc/issues/5): As also stated by @taqtiqa-mark that they need to be implemented
- [ ] installation (https://github.com/justinalsing/dlmmc/issues/3): Your installation instructions are still a bit vague. As a user, I do not want to test whether it works with `pip` or not, I simply want to install it and it should work
- [ ] references: thanks for implementing the detailed references. As noted by your colleague @pankajkarman (https://github.com/justinalsing/dlmmc/issues/4), they have to be reviewed.

If you work on these issues, please directly comment in the issues in your repository as this is also recommended in the review process guidelines [above](https://github.com/openjournals/joss-reviews/issues/1157#issue-396558060)

> Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the [target repository](https://github.com/justinalsing/dlmmc) and link to those issues (especially acceptance-blockers) in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)
