### 2019-02-07 justinalsing/dlm#5 @taqtiqa-mark
I wonder if it isn't worth breaking this issue into three?  To be completed in this order.
 - Vanilla Test suite.
 - Parallel test suite.
 - CI build that executes the above.

@justinalsing, just a suggestion based on personal experience:
 Each of these can be non-trivial to do well, and all the moving parts can be overwhelming if you try to tackle them together.
Also be aware that CI build centric iterations can be much slower than running tests on your local machine.  Generally the CI build of a branch involves all tests and is triggered by a commit that merges several changes from a scratch branch.... on your localhost you'll likely be running subsets of tests with regularity.
