### 2019-02-10 #1157 @danielskatz
Re: automated tests - The JOSS requirement is just "Are there automated tests or manual steps described so that the function of the software can be verified?" While automated tests are generally better for the software, manual steps are sufficient for JOSS acceptance.  Specifically, adding good automated tests can be a longer-term activity than JOSS acceptance.
