### 2019-02-15 #1157 @Chilipp
Thanks @danielskatz for your clarification on this topic. @justinalsing states above (https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-459128351)

> In my mind the tutorial also serves this purpose: if the models compile and the tutorial runs through without error, you’re in good shape. I’ve added a note to the README that the tutorial also serves this purpose.

Would this be sufficient for an acceptance in JOSS? Because the feeling of @taqtiqa-mark and me is that there should be some more tests implemented (see https://github.com/justinalsing/dlmmc/issues/5)
