### 2019-02-15 625469f
fixed references in model descriptions pdf
