### 2019-02-15 justinalsing/dlm#4 @justinalsing
Thank you for pointing this out - I have now corrected this discrepancy in model_descriptions.pdf
