### 2019-02-15 47d4509
Update README.md
Edited README to remove mentions of pip, and added a note regarding which platform the code works on. This hopefully addresses issue #3 https://github.com/justinalsing/dlmmc/issues/3
