### 2019-02-15 justinalsing/dlm#3 @justinalsing
Hi @Chilipp !

Thanks for this - I just edited the README to remove mention of pip as requested and added a note on which platforms the code runs on (Windows, Linux and Mac).

You're right there are some limitations to the functionality of pystan on Windows, but I have close collaborators who are using the code extensively on windows so these limitations don't appear to impact the use of the code. I believe the restrictions are isolated to the way pystan implements parallel processing on Windows, which is not integral to the functioning of the dlmmc code.

Thanks!
Justin
