### 2019-02-15 4ed4f7f
removed pycache and DS_store files (attempt 2)
