### 2019-02-18 #1157 @Chilipp
Hey @danielskatz and @justinalsing. Sorry. Sure, if that is sufficient for the criteria of JOSS, I won't make a stand against it
