### 2019-02-27 #1157 @danielskatz
👋 @taqtiqa-mark - can you let us know if this is acceptable to you?

> ... to beef up the tutorial to (a) have sections to run all of the implemented models rather than just the "vanilla" one, and (b) run each model on mock data generated from the same model so that the examples serve as explicit validation tests in addition to testing whether the code runs.

In my opinion, this would meet the JOSS  criteria.

Also, how is your review coming otherwise?
