### 2019-02-28 #1157 @justinalsing
I plan on proceeding with the above plan anyway (and can add more later if needed), just need to find time to sit down and do this, hopefully next week (maybe sooner)
