### 2019-03-11 7b3b5b2
added validation test notebook
