### 2019-03-11 #1157 @justinalsing
Hi All,

Following the discussion above about automated tests, I've included an additional notebook with validation tests for each of the DLM models implemented in the package - I've given some more details in https://github.com/justinalsing/dlmmc/issues/5

Hope this does the trick!

Thanks,
Justin

PS @afron, above it looks like the two reviewers were unassigned when Daniel transferred this to you as editor? Does this need fixing? (thanks!)
