### 2019-03-12 #1157 @arfon
> PS @afron, above it looks like the two reviewers were unassigned when Daniel transferred this to you as editor? Does this need fixing? (thanks!)

Yes, I think that might be a bug.

@Chilipp, @taqtiqa-mark - please come and take another look at this submission following @justinalsing's updates.
