### 2019-03-13 justinalsing/dlm#5 @Chilipp
Thanks @justinalsing ! I just started and apparently there is a file missing: ``utils/dlm.py``

```python
# Import required modules
import pystan
import matplotlib.pyplot as plt
import numpy as np
import time
import scipy.interpolate as interpolate
import netCDF4
import pickle
import scipy.stats as stats
from utils.utils import *
from utils.dlm import *
%matplotlib inline
```

    <ipython-input-1-890c5b3f6cfd> in <module>
          9 import scipy.stats as stats
         10 from utils.utils import *
    ---> 11 from utils.dlm import *
         12 get_ipython().run_line_magic('matplotlib', 'inline')

    ModuleNotFoundError: No module named 'utils.dlm'
