### 2019-03-14 justinalsing/dlm#5 @Chilipp
Thank you @justinalsing. Unfortunately it it seems like I cannot run the test 2 for dlm_vanilla_ar2 because of the following error. Can you confirm this error?

```

    RuntimeError                              Traceback (most recent call last)

    <ipython-input-4-c33d4d85b5cc> in <module>
         64
         65 # Ok, let's run it
    ---> 66 fit = dlm_model.sampling(data=input_data, iter=4000, warmup=1000, chains=1, init = [initial_state], verbose=True, pars=('sigma_trend', 'sigma_seas', 'sigma_AR', 'rhoAR1', 'rhoAR2', 'trend', 'slope', 'beta', 'seasonal', 'ar'))
         67
         68 # Plot the results:

    ~/miniconda/envs/dlmmc/lib/python3.7/site-packages/pystan/model.py in sampling(self, data, pars, chains, iter, warmup, thin, seed, init, sample_file, diagnostic_file, verbose, algorithm, control, n_jobs, **kwargs)
        670
        671         seed = pystan.misc._check_seed(seed)
    --> 672         fit = self.fit_class(data, seed)
        673
        674         m_pars = fit._get_param_names()

    stanfit4anon_model_3780e43ed9ee465806912a94b90d6e06_7690838862458284755.pyx in stanfit4anon_model_3780e43ed9ee465806912a94b90d6e06_7690838862458284755.StanFit4Model.__cinit__()

    RuntimeError: Exception: variable does not exist; processing stage=data initialization; variable name=nx; base type=int  (in 'unknown file name' at line 6)

```

Here is some information about the working environment:

<details><summary>conda list</summary>

```
# packages in environment at /Users/psommer/miniconda/envs/dlmmc:
#
# Name                    Version                   Build  Channel
appnope                   0.1.0                 py37_1000    conda-forge
attrs                     19.1.0                     py_0    conda-forge
backcall                  0.1.0                      py_0    conda-forge
blas                      1.1                    openblas    conda-forge
bleach                    3.1.0                      py_0    conda-forge
bzip2                     1.0.6             h1de35cc_1002    conda-forge
ca-certificates           2019.3.9             hecc5488_0    conda-forge
certifi                   2019.3.9                 py37_0    conda-forge
cftime                    1.0.3.4         py37h917ab60_1000    conda-forge
curl                      7.64.0               heae2a1f_2    conda-forge
cycler                    0.10.0                     py_1    conda-forge
cython                    0.29.6           py37h0a44026_0    conda-forge
decorator                 4.3.2                      py_0    conda-forge
defusedxml                0.5.0                      py_1    conda-forge
entrypoints               0.3                   py37_1000    conda-forge
freetype                  2.9.1             h597ad8a_1005    conda-forge
hdf4                      4.2.13            hf3c6af0_1002    conda-forge
hdf5                      1.10.4          nompi_h646315f_1105    conda-forge
icu                       58.2              h0a44026_1000    conda-forge
ipykernel                 5.1.0           py37h24bf2e0_1002    conda-forge
ipython                   7.3.0            py37h24bf2e0_0    conda-forge
ipython_genutils          0.2.0                      py_1    conda-forge
ipywidgets                7.4.2                      py_0    conda-forge
jedi                      0.13.3                   py37_0    conda-forge
jinja2                    2.10                       py_1    conda-forge
jpeg                      9c                h1de35cc_1001    conda-forge
jsonschema                3.0.1                    py37_0    conda-forge
jupyter                   1.0.0                      py_1    conda-forge
jupyter_client            5.2.4                      py_3    conda-forge
jupyter_console           6.0.0                      py_0    conda-forge
jupyter_core              4.4.0                      py_0    conda-forge
kiwisolver                1.0.1           py37h04f5b5a_1002    conda-forge
krb5                      1.16.3            hcfa6398_1001    conda-forge
libcurl                   7.64.0               he376013_2    conda-forge
libcxx                    7.0.0                h2d50403_1    conda-forge
libedit                   3.1.20170329      hcfe32e1_1001    conda-forge
libffi                    3.2.1             h0a44026_1005    conda-forge
libgfortran               3.0.1                         0    conda-forge
libnetcdf                 4.6.2             h6b88ef6_1001    conda-forge
libpng                    1.6.36            ha441bb4_1000    conda-forge
libsodium                 1.0.16            h1de35cc_1001    conda-forge
libssh2                   1.8.0             hb1dc21d_1004    conda-forge
llvm-meta                 7.0.0                         0    conda-forge
markupsafe                1.1.1            py37h1de35cc_0    conda-forge
matplotlib                3.0.3                    py37_0    conda-forge
matplotlib-base           3.0.3            py37hf043ca5_0    conda-forge
mistune                   0.8.4           py37h1de35cc_1000    conda-forge
mpi                       1.0                     openmpi    conda-forge
mpi4py                    3.0.1            py37h27a7d74_0    conda-forge
nbconvert                 5.4.1                      py_2    conda-forge
nbformat                  4.4.0                      py_1    conda-forge
ncurses                   6.1               h0a44026_1002    conda-forge
netcdf4                   1.4.3.2          py37hf8bc7f3_0    conda-forge
notebook                  5.7.6                    py37_0    conda-forge
numpy                     1.16.2          py37_blas_openblash486cb9f_0  [blas_openblas]  conda-forge
openblas                  0.3.3             hdc02c5d_1001    conda-forge
openmpi                   3.1.3             h71abe9c_1001    conda-forge
openssl                   1.1.1b               h1de35cc_1    conda-forge
pandoc                    2.6                           1    conda-forge
pandocfilters             1.4.2                      py_1    conda-forge
parso                     0.3.4                      py_0    conda-forge
pexpect                   4.6.0                 py37_1000    conda-forge
pickleshare               0.7.5                 py37_1000    conda-forge
prometheus_client         0.6.0                      py_0    conda-forge
prompt_toolkit            2.0.9                      py_0    conda-forge
ptyprocess                0.6.0                 py37_1000    conda-forge
pygments                  2.3.1                      py_0    conda-forge
pyparsing                 2.3.1                      py_0    conda-forge
pyqt                      5.6.0           py37hc26a216_1008    conda-forge
pyrsistent                0.14.11          py37h1de35cc_0    conda-forge
pystan                    2.17.1.0         py37hf8a1672_2    conda-forge
python                    3.7.1             hbdd33cc_1002    conda-forge
python-dateutil           2.8.0                      py_0    conda-forge
pyzmq                     18.0.1           py37h4cc6ddd_0    conda-forge
qt                        5.6.2             h822fa55_1013    conda-forge
qtconsole                 4.4.3                      py_0    conda-forge
readline                  7.0               hcfe32e1_1001    conda-forge
scipy                     1.2.1           py37_blas_openblash486cb9f_0  [blas_openblas]  conda-forge
send2trash                1.5.0                      py_0    conda-forge
setuptools                40.8.0                   py37_0    conda-forge
sip                       4.18.1          py37h0a44026_1000    conda-forge
six                       1.12.0                py37_1000    conda-forge
sqlite                    3.26.0            h1765d9f_1001    conda-forge
terminado                 0.8.1                 py37_1001    conda-forge
testpath                  0.4.2                 py37_1000    conda-forge
tk                        8.6.9             ha441bb4_1000    conda-forge
tornado                   6.0.1            py37h1de35cc_0    conda-forge
tqdm                      4.31.1                     py_0    conda-forge
traitlets                 4.3.2                 py37_1000    conda-forge
wcwidth                   0.1.7                      py_1    conda-forge
webencodings              0.5.1                      py_1    conda-forge
widgetsnbextension        3.4.2                 py37_1000    conda-forge
xz                        5.2.4             h1de35cc_1001    conda-forge
zeromq                    4.2.5             h0a44026_1006    conda-forge
zlib                      1.2.11            h1de35cc_1004    conda-forge
```

<details><summary>conda info</summary>

```

     active environment : dlmmc
    active env location : /Users/psommer/miniconda/envs/dlmmc
            shell level : 2
       user config file : /Users/psommer/.condarc
 populated config files : /Users/psommer/.condarc
          conda version : 4.6.7
    conda-build version : 3.17.8
         python version : 3.7.1.final.0
       base environment : /Users/psommer/miniconda  (writable)
           channel URLs : https://conda.anaconda.org/conda-forge/osx-64
                          https://conda.anaconda.org/conda-forge/noarch
                          https://repo.anaconda.com/pkgs/main/osx-64
                          https://repo.anaconda.com/pkgs/main/noarch
                          https://repo.anaconda.com/pkgs/free/osx-64
                          https://repo.anaconda.com/pkgs/free/noarch
                          https://repo.anaconda.com/pkgs/r/osx-64
                          https://repo.anaconda.com/pkgs/r/noarch
          package cache : /Users/psommer/miniconda/pkgs
                          /Users/psommer/.conda/pkgs
       envs directories : /Users/psommer/miniconda/envs
                          /Users/psommer/.conda/envs
               platform : osx-64
             user-agent : conda/4.6.7 requests/2.21.0 CPython/3.7.1 Darwin/17.7.0 OSX/10.13.6
                UID:GID : 502:20
             netrc file : /Users/psommer/.netrc
           offline mode : False

```
