### 2019-03-14 0b4f355
updated AR2 model
