### 2019-03-22 justinalsing/dlm#5 @taqtiqa-mark
Hi @justinalsing, I'll likely get time some evening to look at this in the coming week.  Do I understand correctly that you've done something like:

1. Started a new VM (or container)
2. Followed your instructions to setup the software stack.
3. Followed your instructions to run your examples.
4. Observed no errors or missing steps.

Correct?
