### 2019-03-22 justinalsing/dlm#5 @Chilipp
Dear @justinalsing,

sorry for being so slow in this. I ran through your tests and they run through. However, there was another issue with the filenames of your regressors in the notebooks. Please note that UNIX filesystems are case-sensitive, there either please change the filenames of your regressors or change them in the notebook.

As an example: your file name is [regressors/ENSO_MEI_1950_201802.txt](https://github.com/justinalsing/dlmmc/blob/master/regressors/ENSO_MEI_1950_201802.txt) but in your notebook you use [regressors/enso_mei_1950_201802.txt](https://nbviewer.jupyter.org/github/justinalsing/dlmmc/blob/master/dlm_validation_tests.ipynb#Set-time-axis-and-import-the-regressors-to-be-used-in-the-tests). The same happened in the dlm_tutorial notebook.

But I have to say that I like your tests and they are really useful to judge the quality of dlmmc.
