### 2019-03-22 justinalsing/dlm#5 @justinalsing
Hi @Chilipp - not a problem at all, and thanks for the comment. I've now made the regressor filenames case consistent with the actual filenames in the notebook examples (and pushed the changes)

Thanks
