### 2019-03-28 #1157 @taqtiqa-mark
@danielskatz, In light of the feedback I'd say I'm substantively finished the review.  I'm happy with the open tickets tracking the outstanding issues.

I'm about to work through the installation instructions and run the examples-as-test-suite... which raise a question:
**Rather than installation instructions etc, would JOSS accept a link(s) VM image(s) that is publicly available?** e.g. https://www.osboxes.org/virtualbox-images/

The reason I ask is, the author has indicated he had done something like start with a vanilla VM and tested his that his installation was able to run the examples.
My experience tells me that the current installation/configuration instructions are no where near enough to get a new user up and running without considerable trial and error and likely opening issues - not ideal given the audience for this publication are domain SMEs.

If it was acceptable for the author to provide his VM I expect we are all going to be saved considerable time and effort - particularly w.r.t. to the MPI setup.
