### 2019-03-28 #1157 @justinalsing
Hi @taqtiqa-mark - super, thanks for the message. Regarding the VM - looking back at the thread, maybe I answered in haste earlier. All I did at that point was re-clone the repo and run everything through (including re-compiling the models, but not re-installing dependencies), I didn't do this in a VM.

I could definitely try what you suggest, though - I guess I'd have to do this for a bunch of different operating systems?

I see the two installation hurdles that have the potential to cause issues as pystan and mpi4py (although I reiterate that the MPI template example I included is really intended as a "bonus" to help people scale-up their runs, rather than an integral part of the package/suite of DLM models). I guess I'm not sure I'll be able to (realistically) provide more comprehensive support for stan and mpi4py than is already provided by the vendors of those packages? But if you think I can, I'll all ears
