### 2019-03-28 81bf3a4
made regressor filenames case consistent in MPI example script
