### 2019-03-28 #1157 @taqtiqa-mark
>  I guess I'd have to do this for a bunch of different operating systems?

This is up to you.  What OS users will you support? What OS's have you tested this with?

I think it is reasonable to expect that you have tried your instructions starting from a "clean" system.

A VM is the most natural and easiest way to get a clean system.  Please record or link to the VM image you start from.

You could try to nuke your bare-metal installation, but in my experience you tend to forget what you'd done in the past to make things work - I think this will end up being more work for you.

Please copy and paste all commands used, **this will be your install script/instructions**.
Once you get it working for one OS anyone can adapt/extend your script for their environment.

>  ... the MPI template example I included is really intended as a "bonus" to help people scale-up their runs, rather than an integral part of the package/suite of DLM models).

Okay, I have no horse in this race, so you need to decide:  Are you willing to support MPI users or not?
