### 2019-04-07 #1157 @danielskatz
👋 @taqtiqa-mark - I wonder if there are some more items in your review checklist that you can check off? This would help make clear what the blocking items are.
