### 2019-04-08 #1157 @taqtiqa-mark
Checklist updated.
Blocking items (Given the last comment I'm assuming the MPI/parallel use cases are out of scope).

- Complete installation instructions (excl. MPI)
