### 2019-04-13 #1157 @justinalsing
Hi @taqtiqa-mark -

I haven't had time to set up the VM etc yet (I had a go but had some trouble getting VirtualBox working on my machine and ran out of time). However, I am going to help some new users get installed and working with the code this week or next - is there anything I could provide you from this process that would satisfy your last checkpoint re validating the installation instructions?

Best,
Justin
