### 2019-04-22 justinalsing/dlmmc#7 @taqtiqa-mark
Clarify MPI status
Blocking openjournals/joss-reviews/issues/1157

Either remove the reference to MPI support, or provide instructions and test cases to validate MPI setup.

Relevant discussion [here](https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-477475710).
