### 2019-04-22 justinalsing/dlmmc#8 @taqtiqa-mark
OS versions: Conda and/or Pip
Blocking openjournals/joss-reviews/issues/1157

I think @Chilipp's [comment is reasonable](https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-461463100):
> As a user, I do not want to test whether it works with pip or not, I simply want to install it and it should work

There is an apparent issue with the install instructions working with `conda` for some users (see #3) and sometimes with `pip` with others (see #6).  Also see [here](https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-483218852):

>One point of contention though: I used pip to install dependencies (which works fine for me and other colleagues who use pystan extensively). However, @Chilipp did not get a working pystan set up when he did pip install, but he did with conda, so suggested that I change the install instructions to conda (which I duly did). But I don't use conda personally or have it set up

There are three OS that the application claims to work on:
1. Windows (all versions?),
1. MacOS (all versions?),
1. Linux (all distributions and all versions?)
  a. Draft Ubuntu 16.04.6 instructions are in #6

Yet there is ambiguity about which installation method should be used `conda` or `pip`: **Please clarify**
