### 2019-04-22 justinalsing/dlm#6 @taqtiqa-mark
If you do signup to [DigitalOcean (referal-link)](https://m.do.co/c/9a152ce8c79e) please signup using the referal link - its some (small) compensation for preparing these instructions for you ;)

[dlmmc-digiticalocean-chef_zero.zip](https://github.com/justinalsing/dlmmc/files/3102622/dlmmc-digiticalocean-chef_zero.zip)
