### 2019-05-01 #1157 @danielskatz
Hi @taqtiqa-mark and @justinalsing,

Just to make sure I understand where things are, it looks like Mark has created three issues

- https://github.com/justinalsing/dlmmc/issues/6
- https://github.com/justinalsing/dlmmc/issues/7
- https://github.com/justinalsing/dlmmc/issues/8

that are blocking his acceptance of the installation part of the submission, which is the last part that needs to be resolved.

And this is now in Justin's court to respond to them.

Is this correct?

I also see https://github.com/justinalsing/dlmmc/issues/5 is still open, but apparently this is not blocking anything for Mark.
