### 2019-05-01 justinalsing/dlm#8 @justinalsing
Since `pip install pystan` is not as stable as using `conda install pystan`, as noted on the [pystan readthedocs](https://pystan.readthedocs.io/en/latest/installation_beginner.html) (and also pointed out by @Chilipp as you note above), my install instructions currently explicitly advise use of `conda`
