### 2019-05-01 justinalsing/dlm#7 @justinalsing
I already provide an example script (`dlm_lat_alt_mpi_run.py`) and instructions in the `README` for testing whether the MPI setup works smoothly with the code.

However, as I've discussed multiple times, the MPI example template is included in the Git repo as a bonus to help MPI users (or at least have MPI installed) get going with MPI runs more quickly. The MPI example template is NOT intended as an integral part of the package itself. I therefore feel that this should not be a blocking issue for JOSS acceptance, and it would seem unfortunate to have to remove it since I believe some people will find it useful (as a bonus). @danielskatz do you want to weight in on this?
