### 2019-05-01 justinalsing/dlm#6 @justinalsing
I'm not sure I agree with this issue - you say that "no instructions are provided" for linux, but from your discussion above you successfully installed the code using exactly the installation instructions currently provided in the `README`, ie,

pip3 install numpy scipy matplotlib netCDF4 pystan
python3 compile_stan_models.py

(although I note you used pip rather than conda, which is less reliable for pystan as has been discussed in issue #3)

So it seems that the installation instructions are working fine; I'm not sure what more needs to be done here?
