### 2019-05-01 justinalsing/dlm#8 @taqtiqa-mark
That is great, Conda is fine. I tried to follow your instructions in this [comment](https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-483218852), but ran into issues, specifically *one command in the comment hints that Jupyter needs to be installed, but no mention is made in your `README.md` instructions?*  Apologies if I missed it, but I'm doing this in my spare time and the process so far has not been frictionless.
Specifically:
```
jupyter-nbconvert --to notebook --execute --ExecutePreprocessor.timeout=100000 dlm_tutorial.ipynb >dlm_tutorial.log 2>&1
```

Can you provide conda instructions corresponding to what I provided in issue #6 for pip mentioned in your [comment](https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-483218852)?

Please put the commands showing how to run the example notebook in the readme, and please link to a gist (or some such) that shows a successful run if the notebook does not test for correctness and print a message that all is correct.

This will show that your instructions work on one platform - it doesn't have to be Ubuntu or Linux I'd be happy to see proper instructions for any one platform.
