### 2019-05-01 justinalsing/dlm#7 @taqtiqa-mark
Understood, but you give no instructions on how to obtain the MPI setup that you say works - I don't doubt you but reproducible research is not waving your hand and saying 'it works for me'.

I'd be happy to close this issue if you just cut the MPI sections out into a separate MPI-README.md with a note to the effect that MPI is not a core part of the package and users are responsible for getting a working MPI+Python stack.
