### 2019-05-01 justinalsing/dlm#6 @taqtiqa-mark
Hmm,
> using exactly the installation instructions currently provided in the README, ie,

I don't believe that is a reasonable characterization.  The readme now refers to Conda, which is well and good, but your [comment](https://github.com/openjournals/joss-reviews/issues/1157#issuecomment-483218852) claims `pip` and, on that flavour of Linux you need `pip3`, etc otherwise you are using Python 2.x.

You'll also note you omitted `ipython[all]` which is required to run the example notebook to test the installation is all good - this I gleaned from your comment in the JOSS repository.
However, instructions for running the example (psuedo test suite) is not mentioned in the `README.md`...

\<Hint\> In this issue you have one set OS instructions that are detailed enough to be reproducible, just follow them using equivalent `conda` (`conda3`?) commands and packages on a fresh installation, then cut&paste......\</Hint\>
