### 2019-05-02 justinalsing/dlm#7 @justinalsing
OK - I've changed the `README` to note that MPI is not supported as a central part of this package, pointing interested users instead to a separate file `MPI-README` for a description of how to get the MPI example template up and running (without going as far to provide full support for getting MPI working smoothly for all users on all OSs). I hope this is satisfactory.
