### 2019-05-02 justinalsing/dlm#8 @justinalsing
OK - I've tried to make the installation instructions in the `README` even more explicit (see updated `README`). If users use the Anaconda python distribution (as is now explicitly recommended in the install instructions), then most of the dependencies including ipython and jupyter come pre-baked by default. This is preferred both because of the number of dependencies that come automatically, but also as noted because conda plays better with pystan.

I've included a note indicating that installation using pip3 is possible but done at the users own risk (some users might have reasons for not wanting to use Anaconda, but then they are on their own if they run into issues with pip and pystan, although I provide a link to the pystan readthedocs for advice if they need it).

I have also included instructions for executing the test suite (using jupyter-nbconvert etc) as recommended.

I'm copy-pasting the new install instructions from the updated `README` here for convenience:

------------------------------

### Installation

Once you have downloaded the code from this repository you're ready to install dependencies and get set-up.

The code is python3 and has the following dependencies: [numpy](http://www.numpy.org), [scipy](https://www.scipy.org), [matplotlib](https://matplotlib.org), [jupyter](https://jupyter.org/install), [ipython](https://ipython.org/install.html), [netCDF4](https://pypi.org/project/netcdf/), [pystan](https://pystan.readthedocs.io/en/latest/).

**Installation with conda (recommended)**

The most painless way to get set up is using the [Anaconda python distribution](https://www.anaconda.com/distribution/) (recommended), which comes with most of the dependencies as default. The remaining dependencies can then be installed using `conda install` and the DLM models compiled by running:

```
conda install pystan netCDF4
python3 compile_stan_models.py
```

This second line compiles all of the DLM models on your machine, saves them in `models/`, and then you're ready to start DLMing! Jump straight into the jupyter notebook tutorial `dlm_tutorial.ipynb` (see below), or if you prefer you can run a test suite to check that the install worked and all models run smoothly by executing (this will take some minutes to run through):

```
jupyter-nbconvert --to notebook --execute --ExecutePreprocessor.timeout=100000 dlm_validation_tests.ipynb
```

Finally, if you want to see what a successful installation looks like, see `INSTALL.md`.

**Installation with pip (at your own risk)**

Anaconda is not a _requirement_ for installing dlmmc, but is recommended because it works robustly with pystan. If you would rather use a different python distribution and `pip3` for installing dependencies, you are welcome to (at your own risk); see the [pystan readthedocs](https://pystan.readthedocs.io/en/latest/installation_beginner.html) for advice on installing pystan using `pip3` if you run into problems. Note that if you do not use Anaconda you will also have to install the other dependencies listed above, ie.,

```
pip3 install numpy scipy ipython[all] jupyter matplotlib netCDF4 pystan
python3 compile_stan_models.py
```

**Platforms**

dlmmc has been successfully installed on Mac, Linux and Windows. Note that there are some limitations to the functionality of [pystan on Windows](https://pystan.readthedocs.io/en/latest/windows.html), but these do not restrict the use of the dlmmc package for Windows users.

------------------------------
