### 2019-05-02 #1157 @justinalsing
Hi @taqtiqa-mark - I've done some more work on the install instructions and responded to your outstanding issues. I hope we are getting close now!
