### 2019-05-02 justinalsing/dlm#7 @danielskatz
This seems reasonable to me

@taqtiqa-mark - what do you think?
