### 2019-05-02 #1157 @danielskatz
👋 @justinalsing - can you make a new archive of the software in zenodo, figshare, etc.?  And paste the DOI here?  Also, if the version has changed, please let me know the new version number.
