+++ README.md
@@ -1,3 +1,5 @@
+[![DOI](https://zenodo.org/badge/148547735.svg)](https://zenodo.org/badge/latestdoi/148547735)
+
 # DLMMC
 
 **Dynamical Linear Modelling** (DLM) regression code in python for analysis of time-series data. The code is targeted at atmospheric time-series analysis, with a detailed worked example (and data) included for stratospheric ozone, but is a fairly general suite of state space model that can be applied or extended to a wide range of problems.
